requirements.txt
python-telegram-bot
py-zabbix
os
functools
datetime
pybix
